+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Added string replacement for `processEndMessage` printing the return code of the process.
 
+### Changed
+
+* Changed the default `processEndMessage` to print the return code of the process.
+
 ## [0.0.7]
 
 ### Added