+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Printing of process calls.
 * Added kill command for the process behind the selected document.
 * Added commando to kill all processes running at the moment.
+* Select processes via command and kill them.
 
 ## [0.0.8]