+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Select processes via command and send some input.
 * Added send some input to the process behind the selected document.
 * Added debugging of processes by converting processes to debug sessions using templates.
+* Error handling for processes was added.
 
 ## [0.0.8]