+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 
 * Use the current file editor for scrolling down on processes.
+* All file editors viewing a process will follow down if the last row was selected.
 
 ## [0.0.9]